add MVVM Pattern. add Repository Pattern.
