--- lib/api/model/NewsResponse.dart
@@ -1,99 +1,0 @@
-import 'package:news_app/api/model/SoursesRespons.dart';
-
-/// status : "ok"
-/// totalResults : 6008
-/// articles : [{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Rent cap: Consultation measure for social housing in England announced","description":"The government says it wants to avoid rents rising up to 11% for some of England's poorest families.","url":"https://www.bbc.co.uk/news/uk-62732367","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/DB59/production/_126535165_gettyimages-90678832.jpg","publishedAt":"2022-08-30T23:02:29Z","content":"By Andre Rhoden-PaulBBC News\r\nPeople living in social housing in England could see rent increases capped below inflation next year, the government has said. \r\nIt has launched a consultation on plans … [+2610 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Ricardo Quaresma: Portugal's flamboyant former star and his stand for Roma rights","description":"Well known for his flair play on the football pitch, ex-Portugal winger Ricardo Quaresma is a leading voice in the Roma community.","url":"https://www.bbc.co.uk/sport/football/62687226","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/14483/production/_126457038_quaresma_cp_getty.jpg","publishedAt":"2022-08-30T23:01:52Z","content":"Known for his flamboyant range on the pitch, Quaresma is a powerful influence off it too\r\nA mobile phone was a luxury that a young Ricardo Quaresma could not afford.\r\nIn May 2000 he was away in Israe… [+8563 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Meet Conservative members choosing the next PM","description":"The BBC spoke to party members in Norwich who will vote for the next leader and UK prime minister.","url":"https://www.bbc.co.uk/news/av/uk-politics-62731003","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/F701/production/_126533236_p0cx9ftd.jpg","publishedAt":"2022-08-30T23:01:16Z","content":"Last Soviet leader, Mikhail Gorbachev, dies aged 91. Video, 00:03:07Last Soviet leader, Mikhail Gorbachev, dies aged 91"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Everton: Anthony Gordon transfer deadline has passed, says Frank Lampard","description":"Everton boss Frank Lampard says the deadline for which he would consider selling winger Anthony Gordon had now passed.","url":"https://www.bbc.co.uk/sport/football/62732848","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/173A7/production/_126534159_gettyimages-1419390795.jpg","publishedAt":"2022-08-30T22:47:49Z","content":"Anthony Gordon has now scored two Premier League goals this season, half of his total for the whole of last season\r\nEverton boss Frank Lampard said the deadline for which he would consider selling wi… [+2229 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"When Mikhail Gorbachev visited Margaret Thatcher in Downing Street","description":"Former prime minister, Margaret Thatcher called their 1989 meeting \"deep\" and \"very friendly\".","url":"https://www.bbc.co.uk/news/av/uk-62729200","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/F5C1/production/_126531926_p0cxbt5k.jpg","publishedAt":"2022-08-30T22:20:07Z","content":"In April 1989, then Soviet leader Mikhail Gorbachev met with Prime Minister Margaret Thatcher in 10 Downing Street. Mrs Thatcher hailed the talks with as \"deep\" and \"very friendly\".\r\nMr Gorbachev was… [+234 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Southampton 2-1 Chelsea: Ralph Hasenhuttl praises 'perfect performance'","description":"Southampton boss Ralph Hasenhuttl praises his side's \"perfect performance\" as they beat Chelsea 2-1 to get their first home win of the season.","url":"https://www.bbc.co.uk/sport/av/football/62733182","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/1533B/production/_126534868_p0cxbstv.jpg","publishedAt":"2022-08-30T22:19:10Z","content":"Southampton boss Ralph Hasenhuttl praises his side's \"perfect performance\" as they beat Chelsea 2-1 to get their first home win of the season.\r\nMATCH REPORT: Southampton 2-1 Chelsea\r\nWatch the Premie… [+93 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"In pictures: The life of ex-Soviet leader Mikhail Gorbachev","description":"The ex-Soviet leader - who was one of the 20th century's most influential figures - dies age 91.","url":"https://www.bbc.co.uk/news/world-europe-62732829","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/04E5/production/_126535210_be8a025bec47ea26a4091358d3c754138dec137a.jpg","publishedAt":"2022-08-30T22:18:33Z","content":"Mikhail Gorbachev, the man who would go on to become one of the most influential political figures of the 20th Century, was born into relative poverty on 2 March 1931 in the Stavropol region of south… [+3925 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Chelsea need to 'toughen up', says boss Thomas Tuchel after Saints defeat","description":"Chelsea boss Thomas Tuchel says his side need to \"toughen up and show a different mentality\" after they were beaten by Southampton.","url":"https://www.bbc.co.uk/sport/football/62733061","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/8D9D/production/_126535263_gettyimages-1419411754.jpg","publishedAt":"2022-08-30T22:17:08Z","content":"Chelsea too easy to beat - Tuchel\r\nChelsea boss Thomas Tuchel says his side need to \"toughen up and show a different mentality\" after they were beaten by Southampton at St Mary's.\r\nThe German boss cr… [+2599 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Leeds 1-1 Everton: Frank Lampard delighted with players' efforts in new system","description":"Everton manager Frank Lampard says he is delighted with his players' efforts in a new system that earned them a 1-1 draw with in-form Leeds United.","url":"https://www.bbc.co.uk/sport/av/football/62733325","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/14BCF/production/_126534948_p0cxbq18.jpg","publishedAt":"2022-08-30T22:15:04Z","content":"Everton manager Frank Lampard says he is delighted with his players' efforts in a new system that earned them a 1-1 draw with in-form Leeds United.\r\nMATCH REPORT: Leeds 1-1 Everton\r\nWatch the Premier… [+92 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"2022/08/30 22:00 GMT","description":"The latest five minute news bulletin from BBC World Service.","url":"https://www.bbc.co.uk/programmes/w172ykq7496qvlb","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p060dh18.jpg","publishedAt":"2022-08-30T22:06:00Z","content":"The latest five minute news bulletin from BBC World Service."},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Rationing energy is nothing new for off-grid community","description":"Scoraig in the north west Highlands has long had to make careful use of its electricity supply.","url":"https://www.bbc.co.uk/news/uk-scotland-highlands-islands-62686916","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/66B9/production/_126479262_hughtwo.jpg","publishedAt":"2022-08-30T21:58:53Z","content":"People living in the remote off-grid community of Scoraig are used to making careful use of the wind and solar power that they generate.\r\nThe area can only be reached by a boat journey or a five-mile… [+3452 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Southampton 2-1 Chelsea - Thomas Tuchel says his side are 'too easy to beat'","description":"Chelsea head coach Thomas Tuchel says his side are \"too easy to beat\" after defeat at Southampton.","url":"https://www.bbc.co.uk/sport/av/football/62733177","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/1051B/production/_126534866_p0cxbpvv.jpg","publishedAt":"2022-08-30T21:58:26Z","content":"Chelsea head coach Thomas Tuchel says his side are \"too easy to beat\" after defeat at Southampton.\r\nMATCH REPORT: Southampton 2-1 Chelsea\r\nWatch the Premier League highlights on Match of the Day from… [+49 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"It's a medical emergency simulation - but the adrenaline is real","description":"An immersion room simulating emergency scenarios to train student medics is launched in Dundee.","url":"https://www.bbc.co.uk/news/uk-scotland-tayside-central-62723337","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/2227/production/_126534780_immersion03.jpg","publishedAt":"2022-08-30T21:55:52Z","content":"By Laura GoodwinBBC Scotland\r\nThe emergency room at Ninewells Hospital crackles with activity as student medics tend to a patient rushed in with chest pains. \r\nLights flash, machines beep, and the bu… [+2943 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Grieving father's anger at care of son assessed as 'low risk'","description":"Adam Parry from Glasgow died at the age of 32 after multiple seizures due to alcohol withdrawal.","url":"https://www.bbc.co.uk/news/uk-scotland-glasgow-west-62725351","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/136A0/production/_126502597_4.jpg","publishedAt":"2022-08-30T21:51:34Z","content":"By Mary McCoolBBC Scotland news\r\nA grieving father has voiced his anger that a health board failed to flag his son as being vulnerable one month before he died from alcoholism.\r\nAdam Parry died at th… [+5742 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Cost of living: Charity urges more help over school uniforms","description":"One woman says she \"had to fight\" to get a grant and charities say more needs to be done.","url":"https://www.bbc.co.uk/news/uk-wales-62723581","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/C8E6/production/_126503415_whatsappimage2022-08-30at2.50.17pm.jpg","publishedAt":"2022-08-30T21:49:59Z","content":"More support is needed for families to be able to afford school uniforms amid the cost-of-living crisis, a charity has said.\r\nThe Bevan Foundation said the eligibility criteria for school uniform was… [+6852 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Leeds 1-1 Everton: Jesse Marsch frustrated with visitors time-wasting","description":"Leeds manager Jesse Marsch says he was frustrated the referee did not deal with what he saw as Everton's attempts to waste time in their 1-1 draw.","url":"https://www.bbc.co.uk/sport/av/football/62732959","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/FDAF/production/_126534946_p0cxbny4.jpg","publishedAt":"2022-08-30T21:49:52Z","content":"Leeds manager Jesse Marsch says he was frustrated the referee did not deal with what he saw as Everton's attempts to waste time in their 1-1 draw at Elland Road.\r\nMATCH REPORT: Leeds 1-1 Everton\r\nWat… [+106 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Warning over early deaths of learning disabled","description":"Children with a learning disability in Scotland are more likely to die prematurely, a study finds.","url":"https://www.bbc.co.uk/news/uk-scotland-62725103","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/A65D/production/_126498524_ld2.jpg","publishedAt":"2022-08-30T21:48:44Z","content":"By Ian Hamilton &amp; Georgina HayesBBC Scotland\r\nChildren with a learning disability in Scotland are more likely to die prematurely - often from treatable causes, a report has found.\r\nThe Scottish L… [+4849 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Mikhail Gorbachev: Last Soviet leader aged 91","description":"The hospital where Mikhail Gorbachev died said he had been suffering from a long and serious illness.","url":"https://www.bbc.co.uk/news/av/world-europe-62729199","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/8091/production/_126531923_p0cxbmd6.jpg","publishedAt":"2022-08-30T21:44:05Z","content":"The last Soviet leader, Mikahil Gorbachev, has died at the age of 91. \r\nMr Gorbachev came into power in 1985 and was recognised for opening up the USSR and for his rapprochement with the West, but he… [+177 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Energy prices: Businesses facing unbearable pressures - Vaughan Gething","description":"The economy minister will meet with business leaders to discuss soaring energy costs.","url":"https://www.bbc.co.uk/news/uk-wales-62727373","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/103EB/production/_118093566_vaughangething.jpg","publishedAt":"2022-08-30T21:43:16Z","content":"Businesses are facing \"increasingly unbearable pressures\" because of soaring energy and fuel bills, according to Wales' economy minister.\r\nVaughan Gething is due to meet with businesses leaders on We… [+4763 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"The Hundred: Matthew Wade hits brilliant 81 - best shots","description":"Australia's Matthew Wade makes a brilliant 81 off 42 balls to set up Birmingham Phoenix's one-wicket victory over London Spirit at Lord's.","url":"https://www.bbc.co.uk/sport/av/cricket/62732380","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/2D5D/production/_126531611_p0cxbn9v.jpg","publishedAt":"2022-08-30T21:41:20Z","content":"Australia's Matthew Wade makes a brilliant 81 off 42 balls to set up Birmingham Phoenix's one-wicket victory over London Spirit at Lord's.\r\nCatch up with all The Hundred action on BBC iPlayer.\r\nAvail… [+21 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Fulham 2-1 Brighton: Aleksandar Mitrovic was unbelievable - Marco Silva","description":"Marco Silva reacts to Fulham's win over Brighton as Aleksandar Mitrovic scored his fifth of the season.","url":"https://www.bbc.co.uk/sport/av/football/62732462","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/68DB/production/_126534862_p0cxbmbx.jpg","publishedAt":"2022-08-30T21:40:35Z","content":"Fulham head coach Marco Silva says Aleksandar Mitrovic was \"unbelievable\" as he scored his fifth goal of the season in a 2-1 win against Brighton.\r\nMATCH REPORT: Fulham 2-1 Brighton\r\nWatch the Premie… [+93 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Crystal Palace 1-1 Brentford: Thomas Frank pleased with 'brilliant game of football'","description":"Brentford manager Thomas Frank admits he is disappointed Brentford could only draw with Crystal Palace after missing two good chances to win.","url":"https://www.bbc.co.uk/sport/av/football/62732956","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/AF8F/production/_126534944_p0cxbmhm.jpg","publishedAt":"2022-08-30T21:37:24Z","content":"Brentford manager Thomas Frank admits he is disappointed Brentford could only draw with Crystal Palace after missing two good chances to win.\r\nMATCH REPORT: Crystal Palace 1-1 Brentford\r\nWatch the Pr… [+97 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Tony Mowbray: Sunderland appoint former West Brom and Blackburn manager as boss","description":"Sunderland appoint former Blackburn Rovers boss Tony Mowbray as their new head coach on a two-year deal.","url":"https://www.bbc.co.uk/sport/football/62733033","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/11787/production/_124395517_bbcbreakingnewsgraphic.jpg","publishedAt":"2022-08-30T21:33:00Z","content":"Sunderland have appointed former Blackburn boss Tony Mowbray as their new head coach on a two-year deal.\r\nThe 58-year-old replaces Alex Neil after he left to take over at Championship rivals Stoke Ci… [+233 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Ukraine grain ship docks in Djibouti","description":"The first grain from Ukraine in months reaches the Horn of Africa. It is the first vessel carrying Ukrainian grain to arrive in Africa since Russia invaded the country in February. Tens of millions of people in the region are in need of food aid - a combinati…","url":"https://www.bbc.co.uk/programmes/w172yfc47gzt51c","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p0cxb67n.jpg","publishedAt":"2022-08-30T21:32:00Z","content":"The first grain from Ukraine in months reaches the Horn of Africa. It is the first vessel carrying Ukrainian grain to arrive in Africa since Russia invaded the country in February. Tens of millions o… [+605 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"In Canada fears of 'dangerous' politics mounting","description":"Politicians call for better security after video shows Deputy PM Chrystia Freeland being harrassed.","url":"https://www.bbc.co.uk/news/world-us-canada-62718193","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/4B2B/production/_126534291_gettyimages-913776330.jpg","publishedAt":"2022-08-30T21:27:25Z","content":"By Bernd Debusmann JrBBC News\r\nOnline threats, racist or misogynistic insults, public harassment and outright physical intimidation are just some of the behaviours that officials warn are changing th… [+5813 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Crystal Palace 1-1 Brentford: Patrick Vieira frustrated with defending on equaliser","description":"Crystal Palace manager Patrick Vieira says he was frustrated with the defending that gave Brentford an equaliser in their 1-1 draw.","url":"https://www.bbc.co.uk/sport/av/football/62732953","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/3A5F/production/_126534941_p0cxbjqh.jpg","publishedAt":"2022-08-30T21:26:09Z","content":"Crystal Palace manager Patrick Vieira says he was frustrated with the defending that led to his side conceding and equaliser in their 1-1 draw with Brentford.\r\nMATCH REPORT:Crystal Palace 1-1 Brentfo… [+1 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Annan Athletic 1-4 Aberdeen (AET): Jim Goodwin's side labour past League two opponents","description":"Aberdeen avoided one of the biggest upsets in their history as they laboured past League 2 Annan Athletic in extra-time to reach the Scottish League Cup quarter-finals.","url":"https://www.bbc.co.uk/sport/football/62712674","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/10B5B/production/_126534486_20528874.jpg","publishedAt":"2022-08-30T21:21:53Z","content":"Annan captain Steven Swinglehurst forced extra-time with a powerful header at Galabank\r\nAberdeen avoided one of the biggest upsets in their history as they laboured past League 2 Annan Athletic in ex… [+4533 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Fulham 2-1 Brighton: Graham Potter says Seagulls not good enough at Craven Cottage","description":"Brighton head coach Graham Potter admits his side were not good enough at Craven Cottage as Fulham ended their nine-game unbeaten run in the Premier League.","url":"https://www.bbc.co.uk/sport/av/football/62732457","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/17663/production/_126534859_p0cxbhrf.jpg","publishedAt":"2022-08-30T21:19:05Z","content":"Brighton head coach Graham Potter admits his side were not good enough at Craven Cottage as Fulham ended their nine-game unbeaten run in the Premier League.\r\nMATCH REPORT: Fulham 2-1 Brighton\r\nWatch … [+103 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"2022/08/30 21:00 GMT","description":"The latest five minute news bulletin from BBC World Service.","url":"https://www.bbc.co.uk/programmes/w172ykq7496qqv6","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p060dh18.jpg","publishedAt":"2022-08-30T21:06:00Z","content":"The latest five minute news bulletin from BBC World Service."},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Leeds United 1-1 Everton: Luis Sinisterra scores first Premier League goal for Whites","description":"Luis Sinisterra scores his first Premier League goal for Leeds as they drew with Everton at Elland Road.","url":"https://www.bbc.co.uk/sport/football/62641732","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/1025F/production/_126534166_gettyimages-1242824206.jpg","publishedAt":"2022-08-30T21:03:58Z","content":"Luis Sinisterra joined Leeds in the summer from Feyenoord\r\nLuis Sinisterra scored his first Premier League goal for Leeds as they came from a goal behind to draw with Everton in a pulsating encounter… [+4850 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"The Soviet leader who helped end the Cold War","description":"His efforts to democratise his country led to the fall of communism, the end of the Cold War and the break-up of the Eastern Bloc.","url":"https://www.bbc.co.uk/news/world-europe-13669848","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/DA39/production/_118456855_899d94c4-63cf-4b54-9c71-6b5e28e22a71.jpg","publishedAt":"2022-08-30T21:02:17Z","content":"Mikhail Gorbachev was one of the most influential political figures of the 20th Century.\r\nHe presided over the dissolution of a Soviet Union that had existed for nearly 70 years and had dominated hug… [+7778 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Canada reviews anti-racism funding after 'vile' tweets","description":"The prime minister says its \"unacceptable\" money went to a group whose consultant made anti-Semitic posts.","url":"https://www.bbc.co.uk/news/world-us-canada-62731191","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/124B5/production/_126533947_gettyimages-1242661888.jpg","publishedAt":"2022-08-30T20:57:54Z","content":"Canadian Prime Minister Justin Trudeau has said it is \"unacceptable\" that his government awarded money to an anti-racism group whose consultant made anti-Semitic remarks on social media.\r\nCanada will… [+1937 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Mikhail Gorbachev: Remembering a warm-hearted and generous man","description":"The BBC's Steve Rosenberg recalls his meetings with the late Mikhail Gorbachev over the past 20 years.","url":"https://www.bbc.co.uk/news/world-europe-62062852","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/0E1A/production/_125801630_gorbachevraisaphoto4.jpg","publishedAt":"2022-08-30T20:51:41Z","content":"By Steve RosenbergRussia Editor, Moscow\r\nIt's March 2013 and I'm interviewing Mikhail Gorbachev at his Moscow think tank. After half an hour the former Soviet leader announces 'Vsyo!' ('That's your l… [+7273 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"Bbc News","title":"Last Soviet leader Mikhail Gorbachev dies at 91","description":"The former president opened up the USSR and saw it collapse in 1991.","url":"https://www.bbc.co.uk/news/uk-62732447","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/3518/production/_126529531_mediaitem126529530.jpg","publishedAt":"2022-08-30T20:50:36Z","content":"Former Soviet leader Mikhail Gorbachev has died aged 91.\r\nMr Gorbachev, who took over in 1985, is best known for opening up the USSR and for his rapprochement with the West, but he was unable to prev… [+1017 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Dundee 3-0 Falkirk: Championship side cruise into League Cup quarter-finals","description":"Dundee booked their place in the last eight of the Scottish League Cup with a comfortable win over League 1 Falkirk at Dens Park.","url":"https://www.bbc.co.uk/sport/football/62712675","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/17CA3/production/_126534479_20528763.jpg","publishedAt":"2022-08-30T20:50:29Z","content":"Zach Robinson scored Dundee's third goal from the spot\r\nDundee booked their place in the last eight of the Scottish League Cup with a comfortable win over League 1 Falkirk at Dens Park.\r\nLyall Camero… [+3853 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Rangers 3-1 Queen of the South Arfield scores two in League Cup win","description":"Teenage striker Robbie Ure scores a debut goal before Scott Arfield's double as Rangers beat Queen of the South to reach the last eight in the League Cup.","url":"https://www.bbc.co.uk/sport/football/62712695","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/1714F/production/_126534549_20528617.jpg","publishedAt":"2022-08-30T20:46:49Z","content":"Robbie Ure made the dream start to his Rangers debut, with an early goal\r\nA debut goal from teenage striker Robbie Ure set Rangers on their way to an ultimately comfortable League Cup second-round vi… [+4707 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Shamima Begum: Spy for Canada smuggled schoolgirl to Syria","description":"Mohammed Al Rasheed says he shared her passport details with Canada as UK police hunted for her.","url":"https://www.bbc.co.uk/news/uk-62726954","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/DE91/production/_120577965_shamima2.jpg","publishedAt":"2022-08-30T20:45:33Z","content":"By Josh BakerBBC News\r\nShamima Begum, who fled the UK and joined the Islamic State group, was smuggled into Syria by an intelligence agent for Canada.\r\nFiles seen by the BBC show he claimed to have s… [+5309 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"The Hundred: London Spirit v Birmingham Phoenix - Plays of the day","description":"Watch the best of Tuesday's Hundred action as London Spirit hosted Birmingham Phoenix at Lord's.","url":"https://www.bbc.co.uk/sport/av/cricket/62728451","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/161F5/production/_126531609_p0cxbdfd.jpg","publishedAt":"2022-08-30T20:44:26Z","content":"Watch the best of Tuesday's Hundred action as London Spirit hosted Birmingham Phoenix at Lord's.\r\nCatch up with all The Hundred action on BBC iPlayer.\r\nAvailable to UK users only."},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Sheffield United 4-0 Reading: Anel Ahmedhodzic scores twice as Sheffield United go top","description":"Sheffield United return to the top of the Championship with a clinical 4-0 win over Reading to leapfrog them to the league summit.","url":"https://www.bbc.co.uk/sport/football/62641670","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/6E53/production/_126534282_shutterstock_editorial_13353466s.jpg","publishedAt":"2022-08-30T20:43:42Z","content":"Bosnia international Ahmedhodzic scored his second goal of the campaign for Sheffield United\r\nSheffield United leapfrogged Reading to return to the top of the Championship, with a clinical 4-0 win ov… [+3323 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Sammy Chung: Former Wolverhampton Wanderers boss dies aged 90","description":"Former Wolves manager and Watford striker Sammy Chung dies at the age of 90.","url":"https://www.bbc.co.uk/sport/football/62729341","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/1859/production/_126533260_gettyimages-1314538680.jpg","publishedAt":"2022-08-30T20:13:38Z","content":"Sammy Chung was only the second Anglo-Chinese professional player in English football\r\nFormer Wolves manager Sammy Chung has died at the age of 90.\r\nChung made more than 300 appearances as a player f… [+841 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"A penguin in California is getting new orthopaedic boots","description":"The San Diego Zoo penguin suffers from a chronic foot disease but is being helped by new boots.","url":"https://www.bbc.co.uk/news/av/world-us-canada-62690170","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/158B3/production/_126534288_p0cxbb05.jpg","publishedAt":"2022-08-30T20:12:58Z","content":"Lucas, the four-year-old African penguin who lives at the San Diego Zoo, has bumblefoot, a degenerative foot disease found in birds. \r\nHe's been fitted with special new boots that helps him waddle mo… [+14 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"2022/08/30 20:00 GMT","description":"The latest five minute news bulletin from BBC World Service.","url":"https://www.bbc.co.uk/programmes/w172ykq7496qm32","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p060dh18.jpg","publishedAt":"2022-08-30T20:06:00Z","content":"The latest five minute news bulletin from BBC World Service."},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Snow Patrol's Gary Lightbody awarded freedom of home borough","description":"A ceremony took place at Bangor Castle ahead of Wednesday evening's Snow Patrol concert in the city.","url":"https://www.bbc.co.uk/news/uk-northern-ireland-62708840","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/0FCB/production/_126534040_880afba322daea9b6a93f764c948f12fffd0a92d.jpg","publishedAt":"2022-08-30T20:03:23Z","content":"Ards and North Down Borough Council has awarded Snow Patrol frontman Gary Lightbody the freedom of the borough.\r\nA ceremony took place at Bangor Castle ahead of Wednesday evening's Snow Patrol acoust… [+1044 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Himars missiles and anti-radiation pills: Ukraine round-up","description":"Ukraine says US weaponry is changing the course of the war, but fears grow of a possible nuclear disaster.","url":"https://www.bbc.co.uk/news/world-europe-62727955","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/CBE1/production/_126339125_himars5.png","publishedAt":"2022-08-30T20:01:48Z","content":"More than six months after Russian troops crossed the border, there are continuing indications of a major Ukrainian counter-offensive.\r\nUkraine's military says it has destroyed key Russian control po… [+3794 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"US Open: Carlos Alcaraz reaches second round as Sebastian Baez retires","description":"Spanish teenager Carlos Alcaraz reaches the US Open second round as Argentine Sebastian Baez retires from their match.","url":"https://www.bbc.co.uk/sport/tennis/62720086","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/14715/production/_126533738_gettyimages-1242822460.jpg","publishedAt":"2022-08-30T19:36:28Z","content":"In April, Carlos Alcaraz became the youngest man to reach the top 10 since compatriot Rafael Nadal in 2005\r\n<table><tr><th>US Open 2022</th></tr>\r\n<tr><td>Venue: Flushing Meadows, New York Dates: 29 … [+2409 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Devon Allen: Philadelphia Eagles release Olympic hurdler as they cut squad for NFL season","description":"The Philadelphia Eagles release Olympic hurdler Devon Allen after just over a month with the franchise.","url":"https://www.bbc.co.uk/sport/american-football/62729085","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/EC11/production/_126533406_gettyimages-1242186301.jpg","publishedAt":"2022-08-30T19:34:01Z","content":"Devon Allen caught a 55-yard touchdown pass in a pre-season game against the Cleveland Browns\r\nThe Philadelphia Eagles have released Olympic hurdler Devon Allen after just over a month with the franc… [+636 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"The Hundred: Watch the best shots from Ravi Bopara's 45-run innings","description":"Watch the best shots from Ravi Bopara's 45 runs off 20 balls as he hits five sixes for London Spirit against Birmingham Phoenix in their Hundred match at Lord's.","url":"https://www.bbc.co.uk/sport/av/cricket/62728444","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/794A/production/_126505013_p0cxb3lj.jpg","publishedAt":"2022-08-30T19:33:11Z","content":"Watch the best shots from Ravi Bopara's 45 runs off 20 balls as he hits five sixes for London Spirit against Birmingham Phoenix in their Hundred match at Lord's.\r\nCatch up with all The Hundred action… [+44 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"2022/08/30 19:00 GMT","description":"The latest five minute news bulletin from BBC World Service.","url":"https://www.bbc.co.uk/programmes/w172ykq7496qhby","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p060dh18.jpg","publishedAt":"2022-08-30T19:06:00Z","content":"The latest five minute news bulletin from BBC World Service."},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Scottish government 'cannot offer' more money to end pay dispute","description":"Scotland's Deputy First Minister John Swinney said \"people needed to understand there were finite limits\" on finances.","url":"https://www.bbc.co.uk/news/uk-scotland-62726483","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/1111/production/_126496340_edinburgh2-andrewpicken.jpg","publishedAt":"2022-08-30T18:42:33Z","content":"Media caption, Bin strikes have lead to overflowing bins in Glasgow and many other areas of Scotland too\r\nThe Scottish government \"cannot offer\" any more money to help resolve the ongoing council wor… [+4854 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Queens Park Rangers v Hull City","description":"Live coverage of Tuesday's Championship game between Queens Park Rangers and Hull City.","url":"https://www.bbc.co.uk/sport/football/62641675","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/15C36/production/_90724198_football_getty.jpg","publishedAt":"2022-08-30T18:37:43Z","content":"Offside, Hull City. Alfie Jones tries a through ball, but Óscar Estupiñán is caught offside."},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Burnley v Millwall","description":"Live coverage of Tuesday's Championship game between Burnley and Millwall.","url":"https://www.bbc.co.uk/sport/football/62641676","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/15C36/production/_90724198_football_getty.jpg","publishedAt":"2022-08-30T18:37:19Z","content":"Offside, Burnley. Nathan Tella tries a through ball, but Jay Rodríguez is caught offside."},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Scottish Conservative councillor charged over romance fraud","description":"Andrew Polson, who led East Dunbartonshire council, was suspended by the Scottish Conservative Party in June.","url":"https://www.bbc.co.uk/news/uk-scotland-glasgow-west-62730785","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/ACD8/production/_125384244_polson1.jpg","publishedAt":"2022-08-30T18:19:32Z","content":"A former Scottish council leader has been charged in connection with an alleged romance fraud. \r\nEast Dunbartonshire councillor, Andrew Polson, 50, was previously named in a separate civil court acti… [+793 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Kochava faces legal action over sale of location data","description":"Federal Trade Commission says the data could be used to track visits to reproductive-health clinics.","url":"https://www.bbc.co.uk/news/technology-62654155","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/E491/production/_126531585_locationgettyimages-1056449262.jpg","publishedAt":"2022-08-30T18:07:38Z","content":"By Chris VallanceTechnology reporter\r\nA US consumer watchdog has begun legal action against a firm that allegedly sold location data that some fear could help identify people having abortions.\r\nThe F… [+2741 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"2022/08/30 18:00 GMT","description":"The latest five minute news bulletin from BBC World Service.","url":"https://www.bbc.co.uk/programmes/w172ykq7496qclt","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p060dh18.jpg","publishedAt":"2022-08-30T18:06:00Z","content":"The latest five minute news bulletin from BBC World Service."},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Edinburgh Tram Inquiry: Costs to exceed £13m","description":"The inquiry was established in June 2014 to find out what went wrong with the £776m project.","url":"https://www.bbc.co.uk/news/uk-scotland-edinburgh-east-fife-62727780","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/134F5/production/_85939097_85939094.jpg","publishedAt":"2022-08-30T18:05:19Z","content":"The Edinburgh Tram Inquiry will cost more than £13m, Transport Scotland has projected.\r\nThe inquiry was established in June 2014 to find out why the tram project was £400m over budget and was complet… [+2242 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Ryanair will grow even faster in recession, boss says","description":"People will still fly during a downturn but they will choose low-fare airlines, Michael O'Leary says.","url":"https://www.bbc.co.uk/news/business-62728621","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/13A9C/production/_126504508_gettyimages-1242817792.jpg","publishedAt":"2022-08-30T18:04:32Z","content":"By Katy AustinTransport correspondent\r\nRyanair will grow even faster in the event of an economic downturn in the UK, the company's boss has told the BBC. \r\nMichael O'Leary said people would still fly… [+4342 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"One-Day Cup: Darren Stevens guides Kent to victory over Hampshire in semi-final","description":"Darren Stevens' hits an unbeaten 84 as Kent beat Hampshire to reach the One-Day Cup final.","url":"https://www.bbc.co.uk/sport/cricket/62694805","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/12DB1/production/_126533277_gettyimages-1419368017.jpg","publishedAt":"2022-08-30T17:48:22Z","content":"Darren Stevens should play his last one-day game for Kent in the final\r\n<table>\r\n<tr><td>Royal London One-Day Cup semi-final, Ageas Bowl</td></tr><tr><td>Hampshire 310-9 (50 overs): Gubbins 75; Denly… [+1951 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Edinburgh International Book Festival reports strong ticket sales","description":"Organisers report a shift in booking behaviour but \"an overwhelmingly positive\" audience response.","url":"https://www.bbc.co.uk/news/uk-scotland-edinburgh-east-fife-62728859","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/622D/production/_126533152_bookfestival_audience_eibf.jpg","publishedAt":"2022-08-30T17:40:03Z","content":"By Pauline McLeanBBC Scotland arts correspondent\r\nEdinburgh International Book Festival has announced sales of 100,000 tickets for this year's event.\r\nThe programme included 600 events - a third fewe… [+3737 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"One-Day Cup: Dane Vilas hits century as Lancashire beat Sussex in One-Day Cup semi-final","description":"Dane Vilas' superb hundred helps Lancashire recover from 67-5 to beat Sussex in the One-Day Cup semi-final.","url":"https://www.bbc.co.uk/sport/cricket/62695400","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/AD3E/production/_126505344_gettyimages-1419302694.jpg","publishedAt":"2022-08-30T17:39:18Z","content":"Dane Vilas hit eight fours and six sixes in his innings of 121\r\n<table>\r\n<tr><td>Royal London One-Day Cup semi-final, 1st central County Ground, Hove</td></tr><tr><td>Lancashire 319-8 (50 overs): Vil… [+1771 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Kherson: 'Heavy fighting' as Ukraine seeks to retake Russian-held region","description":"Military experts tell the BBC the battle for the southern region could go on for months.","url":"https://www.bbc.co.uk/news/world-europe-62730439","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/10AD9/production/_126531386_8df62cef-b315-4e57-bcb3-917f1d624b9b.jpg","publishedAt":"2022-08-30T17:39:10Z","content":"By Yaroslav LukovBBC News\r\nFierce battles have been reported as Ukraine tries to retake the Russian-occupied southern Kherson region - but military experts have told the BBC it \"won't happen quickly\"… [+3293 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"US Open: Iga Swiatek beats Jasmine Paolini to reach second round","description":"World number one Iga Swiatek makes light work of her US Open opener with a straight-set win over Jasmine Paolini.","url":"https://www.bbc.co.uk/sport/tennis/62729238","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/13218/production/_126506387_iga_swiatek_reuters2.jpg","publishedAt":"2022-08-30T17:27:41Z","content":"Iga Swiatek reached the fourth round of the US Open in 2021\r\n<table><tr><th>US Open 2022</th></tr>\r\n<tr><td>Venue: Flushing Meadows, New York Dates: 29 August-11 September</td></tr><tr><td>Coverage: … [+845 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Charley Kendall: Lincoln City forward joins Sutton United on loan","description":"Sutton United have signed Lincoln City forward Charley Kendall on loan for the remainder of the season.","url":"https://www.bbc.co.uk/sport/football/62731181","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/94F5/production/_126533183_ck_pa.jpg","publishedAt":"2022-08-30T17:25:21Z","content":"Charley Kendall (right) scored on his debut for Lincoln City, helping them past Doncaster Rovers in the Carabao Cup first round\r\nSutton United have signed Lincoln City forward Charley Kendall on loan… [+448 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Scott Parker: Jurgen Klopp felt for ex-Bournemouth boss after sacking","description":"Liverpool boss Jurgen Klopp says he \"felt\" for Scott Parker who was sacked by Bournemouth after Saturday's 9-0 defeat by Liverpool, four games into the season.","url":"https://www.bbc.co.uk/sport/av/football/62731257","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/ACC9/production/_126533244_p0cx9cvt.jpg","publishedAt":"2022-08-30T17:25:21Z","content":"Liverpool manager Jurgen Klopp says he \"felt\" for Scott Parker who was sacked as Bournemouth boss following Saturday's 9-0 defeat by Liverpool, four games into the season.\r\nREAD MORE: Bournemouth sac… [+18 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Israeli court sentences Gaza aid worker to 12 years in prison","description":"An Israeli court convicts World Vision's Mohammed Halabi of diverting funds to Hamas in June.","url":"https://www.bbc.co.uk/news/world-middle-east-62728958","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/8D25/production/_126533163_mediaitem126533162.jpg","publishedAt":"2022-08-30T17:23:57Z","content":"By David GrittenBBC News\r\nAn Israeli court has sentenced a Palestinian aid worker, convicted of diverting funds to the militant group Hamas, to 12 years in prison.\r\nMohammed Halabi, the head of World… [+2410 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Liverpool boss Jurgen Klopp would like to 'risk more' in transfer market","description":"Jurgen Klopp wants Liverpool to \"risk a bit more\" in the transfer market as they continue to look for a midfielder.","url":"https://www.bbc.co.uk/sport/football/62729660","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/FFAA/production/_126505456_gettyimages-1417944547.jpg","publishedAt":"2022-08-30T17:19:28Z","content":"I felt for Parker after sacking - Klopp\r\nJurgen Klopp wants Liverpool to \"risk a bit more\" in the transfer market as they continue to look for a midfielder.\r\nStriker Darwin Nunez is their only summer… [+1165 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"LIV Golf: Cameron Smith's defection delivers a sickening blow to the PGA Tour","description":"Cameron Smith's departure to LIV Golf is a sickening blow to the PGA Tour and those seeking to preserve the sport's status quo.","url":"https://www.bbc.co.uk/sport/golf/62730763","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/33A7/production/_126532231_gettyimages-1242786831.jpg","publishedAt":"2022-08-30T17:14:02Z","content":"Cameron Smith won six PGA Tour titles, including this year's Players Championship\r\nIt was expected but Cameron Smith's departure to LIV Golf remains a sickening blow to the PGA Tour and those seeking… [+6111 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"2022/08/30 17:00 GMT","description":"The latest five minute news bulletin from BBC World Service.","url":"https://www.bbc.co.uk/programmes/w172ykq7496q7vp","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p060dh18.jpg","publishedAt":"2022-08-30T17:06:00Z","content":"The latest five minute news bulletin from BBC World Service."},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"US Open: Cameron Norrie beats Benoit Paire to reach New York second round","description":"British men's number one Cameron Norrie makes light work of unpredictable Frenchman Benoit Paire as he starts his US Open bid in style.","url":"https://www.bbc.co.uk/sport/tennis/62720093","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/1747F/production/_126495359_norrie.jpg","publishedAt":"2022-08-30T16:52:13Z","content":"Cameron Norrie moved to a career high ranking of ninth in the world on Monday\r\n<table><tr><th>US Open 2022</th></tr>\r\n<tr><td>Venue: Flushing Meadows, New York Dates: 29 August-11 September</td></tr>… [+918 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Wesley Fofana: Brendan Rodgers says transfer uncertainty affecting 'environment and stability' at Leicester","description":"Leicester boss Brendan Rodgers says the uncertainty surrounding Wesley Fofana's future is impacting the \"environment and stability\" at the club.","url":"https://www.bbc.co.uk/sport/football/62729651","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/3C5A/production/_126505451_gettyimages-1242484019.jpg","publishedAt":"2022-08-30T16:49:42Z","content":"Chelsea are set to spend in the region of £70m to sign Wesley Fofana from Leicester City\r\nLeicester boss Brendan Rodgers says Wesley Fofana's protracted move to Chelsea is impacting the \"environment … [+1524 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"New £3.6bn London transport funding deal agreed","description":"It secures the long-term future of keeping the capital's Tube trains, railways, buses and trams running.","url":"https://www.bbc.co.uk/news/uk-england-london-62728974","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/173A3/production/_125593159_screenshot2022-06-23183752.png","publishedAt":"2022-08-30T16:37:54Z","content":"A new £3.6bn government bailout to keep Tube trains, railways, buses and trams running in London has been agreed.\r\nThe package includes almost £1.2bn of upfront funding for Transport for London (TfL)… [+3048 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"NHS Lothian operations to resume by end of week","description":"The health board had previously said it would have to prioritise emergency and urgent surgery for two weeks.","url":"https://www.bbc.co.uk/news/uk-scotland-edinburgh-east-fife-62727774","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/1868A/production/_117687999_eri.jpg","publishedAt":"2022-08-30T16:32:35Z","content":"Surgeries at NHS Lothian will resume by the end of the week after a leak in a unit that sterilises surgical equipment caused the postponement of operations.\r\nThe health board had previously said it w… [+1318 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Olivia Pratt-Korbel: PC scooped up girl and ran, inquest told","description":"Fatally-injured Olivia Pratt-Korbel was rushed to hospital in the back of a police car, an inquest hears.","url":"https://www.bbc.co.uk/news/uk-england-merseyside-62729604","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/6CCD/production/_126435872_oliviapratt1.jpg","publishedAt":"2022-08-30T16:32:11Z","content":"A police officer \"scooped\" up Olivia Pratt-Korbel and ran with her to try and save her life after she was shot at her home, an inquest has heard.\r\nOlivia was fatally shot as her mother tried to stop … [+1290 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Leeds: Footage shows teenagers throwing bike on rail line at Cross Gates","description":"Network Rail says it is seeing a rise in anti-social behaviour by young people at the station.","url":"https://www.bbc.co.uk/news/uk-england-leeds-62724096","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/5910/production/_126500822_de54.jpg","publishedAt":"2022-08-30T16:20:51Z","content":"Media caption, Leeds: Footage shows teenagers throwing bike on rail line at Cross Gates\r\nNetwork Rail has released CCTV footage showing two teenagers throwing a bike onto a railway line near Leeds.\r\n… [+1513 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Vuelta a Espana: Remco Evenepoel powers to victory on individual time trial on stage 10 to extend overall lead","description":"Belgium's Remco Evenepoel powers to victory on the individual time trial on stage 10 as he extends his overall lead in the Vuelta a Espana.","url":"https://www.bbc.co.uk/sport/cycling/62728349","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/11050/production/_126521796_gettyimages-1419350328.jpg","publishedAt":"2022-08-30T16:19:28Z","content":"Remco Evenepoel's victory was his first stage win at a Grand Tour\r\nRemco Evenepoel powered to victory on the individual time trial on stage 10 to extend his overall lead in the Vuelta a Espana.\r\nThe … [+3381 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"The Hundred: Megan Schutt takes two wickets in two balls","description":"Watch London Spirit's Megan Schutt take the wickets of Birmingham Phoenix's Amy Jones and Ellyse Perry in consecutive deliveries in their Hundred match at Lord's.","url":"https://www.bbc.co.uk/sport/av/cricket/62728442","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/15FC2/production/_126505009_p0cx8gbh.jpg","publishedAt":"2022-08-30T16:18:02Z","content":"Watch London Spirit's Megan Schutt take the wickets of Birmingham Phoenix's Amy Jones and Ellyse Perry in consecutive deliveries in their Hundred match at Lord's.\r\nCatch up with all The Hundred actio… [+45 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"The Hundred: Naomi Dattani and Alice Monaghan combine for relay catch","description":"Watch London Spirit's Naomi Dattani and Alice Monaghan team up for a \"brilliant\" relay catch to dismiss Birmingham Phoenix's Eva Jones in their Hundred match at Lord's.","url":"https://www.bbc.co.uk/sport/av/cricket/62728443","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/2B2A/production/_126505011_p0cx8r4z.jpg","publishedAt":"2022-08-30T16:17:05Z","content":"Watch London Spirit's Naomi Dattani and Alice Monaghan team up for a \"brilliant\" relay catch to dismiss Birmingham Phoenix's Eva Jones in their Hundred match at Lord's.\r\nCatch up with all The Hundred… [+51 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Marta Temido: Portugal health minister quits after pregnant tourist dies","description":"Portugal's Marta Temido was criticised for closing some natal units, forcing risky transfers for women.","url":"https://www.bbc.co.uk/news/world-europe-62727464","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/1C79/production/_126498270_gettyimages-1235395647.jpg","publishedAt":"2022-08-30T16:15:36Z","content":"By Leo SandsBBC News\r\nPortugal's health minister has resigned hours after reports emerged that a pregnant tourist had died after being turned away from a full maternity ward.\r\nThe 34-year-old Indian … [+2401 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Kenny Shiels: Northern Ireland manager wants to stay 'unless someone thinks any different'","description":"Kenny Shiels reiterates his desire to continue as Northern Ireland manager into the Euro 2025 qualification campaign next year.","url":"https://www.bbc.co.uk/sport/football/62729293","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/79AE/production/_126505113_pe_00913358.jpg","publishedAt":"2022-08-30T16:06:04Z","content":"Kenny Shiels guided Northern Ireland to their major women's tournament in July\r\nKenny Shiels has reiterated his desire to continue as Northern Ireland manager into the Euro 2025 qualification campaig… [+2211 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"2022/08/30 16:00 GMT","description":"The latest five minute news bulletin from BBC World Service.","url":"https://www.bbc.co.uk/programmes/w172ykq7496q43k","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p060dh18.jpg","publishedAt":"2022-08-30T16:06:00Z","content":"The latest five minute news bulletin from BBC World Service."},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Channel migrants: Home Office pays firms to pick up people in small boats","description":"More than £2.5m has been spent this year on firms which \"collect\" people seeking to enter the UK.","url":"https://www.bbc.co.uk/news/uk-england-kent-62728398","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/688A/production/_126026762_tv075313710.jpg","publishedAt":"2022-08-30T16:04:38Z","content":"The government has paid private companies more than £2.5m this year to help pick up people attempting to cross the Channel on small boats.\r\nThe measure, first reported in The Guardian, comes after cr… [+1726 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Creamfields North: South Wales woman dies after falling ill","description":"A 25-year-old woman dies after becoming unwell at Creamfields North music festival in Cheshire.","url":"https://www.bbc.co.uk/news/uk-england-merseyside-62728567","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/148A2/production/_126503148_mediaitem126503147.jpg","publishedAt":"2022-08-30T16:00:08Z","content":"A woman has died days after falling ill at Creamfields North, police have said.\r\nThe 25-year-old was attending the popular music festival in Daresbury when she became unwell on Friday and was taken t… [+1140 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"HRT tsar returns to role overseeing Covid boosters","description":"Madelaine McTernan was the head of the HRT Taskforce, to ease UK supply issues, for four months.","url":"https://www.bbc.co.uk/news/health-62721742","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/12667/production/_124076357_tablets.jpg","publishedAt":"2022-08-30T15:51:50Z","content":"Madelaine McTernan, who was recently appointed hormone-replacement therapy (HRT) tsar, to help deal with UK supply issues, has returned to her role overseeing vaccines, ahead of the planned autumn Co… [+1584 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Books and bomb shelters: Ukraine returns to school","description":"The war in Ukraine has ravaged some schools in the country, with more than 280 totally destroyed.","url":"https://www.bbc.co.uk/news/av/world-europe-62729196","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/1C52/production/_126505270_p0cx8snp.jpg","publishedAt":"2022-08-30T15:45:09Z","content":"As the new school year approaches, Ukraine has been preparing for students to return to education. \r\nSince Russia's invasion of the country began on 24 February, more than 280 schools have been total… [+215 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Spurs defender Reguilon loaned to Atletico Madrid","description":"Tottenham's Spain international defender Sergio Reguilon joins Atletico Madrid on a season-long loan.","url":"https://www.bbc.co.uk/sport/football/62729334","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/D12E/production/_126505535_gettyimages-1369747603.jpg","publishedAt":"2022-08-30T15:38:38Z","content":"Sergio Reguilon has played 67 times in all competitions in his two seasons at Tottenham.\r\nTottenham defender Sergio Reguilon has joined Atletico Madrid on a season-long loan.\r\nThe Spain international… [+349 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Eurostar to suspend direct trains to Disneyland Paris","description":"The service from London will be suspended from 5 June next year to focus on \"core routes\".","url":"https://www.bbc.co.uk/news/business-62728618","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/CAD0/production/_126502915_minniemouse.png","publishedAt":"2022-08-30T15:35:45Z","content":"Eurostar is to suspend its direct train service between London and Disneyland Paris. \r\nThe company said the route would be axed from 5 June next year for the duration of the summer. \r\nIt said the dec… [+2872 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Inventing Anna: Rachel DeLoache Williams sues Netflix for defamation","description":"Rachel DeLoache Williams says the true-life series caused \"catastrophic damage\" to her reputation.","url":"https://www.bbc.co.uk/news/entertainment-arts-62727106","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/2666/production/_126503890_inventinganna_netflix.jpg","publishedAt":"2022-08-30T15:27:51Z","content":"A former Vanity Fair photo editor has sued Netflix over how she was portrayed in the hit series Inventing Anna.\r\nRachel DeLoache Williams was among the real people who were depicted in the true-life … [+2427 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"The powerful cleric behind Iraq protests","description":"Moqtada al-Sadr has been a powerful influence in Iraq since the fall of Saddam Hussein in 2003.","url":"https://www.bbc.co.uk/news/world-middle-east-12135160","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/4736/production/_126503281_mediaitem126503015.jpg","publishedAt":"2022-08-30T15:18:28Z","content":"Moqtada al-Sadr has been a powerful figure in Iraq since the fall of Saddam Hussein in 2003.\r\nThe Shia Muslim cleric rose to prominence after the US-led invasion, when his fearsome Mehdi Army militia… [+8223 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Queens of crime: Val McDermid reveals legal threat over Agatha Christie title","description":"Val McDermid says Agatha Christie's estate has objected to her being billed as the \"queen of crime\".","url":"https://www.bbc.co.uk/news/entertainment-arts-62727105","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/E5CE/production/_126503885_mcdermid_bbc1.jpg","publishedAt":"2022-08-30T15:15:28Z","content":"Author Val McDermid has said Agatha Christie's estate has threatened legal action against her publisher for promoting her as the \"queen of crime\".\r\nThat phrase is trademarked by the descendants of Ch… [+3533 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Tomato truck crash causes huge mess on California highway","description":"Over 150,000 tomatoes spilled across a busy highway causing seven cars to crash.","url":"https://www.bbc.co.uk/news/av/world-us-canada-62729166","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/102CA/production/_126505266_p0cx8lcq.jpg","publishedAt":"2022-08-30T15:13:28Z","content":"A truck has spilled more than 150,000 tomatoes across a busy motorway in the US state of California, causing seven cars to crash.\r\nCars drove over the squelching tomatoes, creating a dangerous slurry… [+142 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Manuel Akanji: Manchester City in talks with Borussia Dortmund over defender","description":"Manchester City are in advanced talks with Borussia Dortmund about signing Switzerland defender Manuel Akanji.","url":"https://www.bbc.co.uk/sport/football/62728209","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/7C60/production/_126504813_gettyimages-1240202280.jpg","publishedAt":"2022-08-30T15:06:09Z","content":"Manuel Akanji has yet to play for Dortmund this season\r\nManchester City are in advanced talks with Borussia Dortmund about signing Switzerland defender Manuel Akanji.\r\nNo deal for Akanji has been agr… [+440 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"2022/08/30 15:00 GMT","description":"The latest five minute news bulletin from BBC World Service.","url":"https://www.bbc.co.uk/programmes/w172ykq7496q0cf","urlToImage":"https://ichef.bbci.co.uk/images/ic/1200x675/p060dh18.jpg","publishedAt":"2022-08-30T15:06:00Z","content":"The latest five minute news bulletin from BBC World Service."},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Jackson, Mississippi has 'no water to drink or flush toilets'","description":"The 180,000 residents of Jackson, Mississippi \"indefinitely\" have no reliable running water.","url":"https://www.bbc.co.uk/news/world-us-canada-62727586","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/ECE0/production/_126504606_4806266425e1f93ec01cbe8ad9249da70aff6026.jpg","publishedAt":"2022-08-30T15:05:37Z","content":"By Sam CabralBBC News, Washington\r\nSome 180,000 residents in Jackson, Mississippi have \"indefinitely\" lost access to reliable running water after excessive rainfall and flooding.\r\nRising floodwaters … [+2564 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Nick Robinson: Liz Truss cancelled, but public want their leaders interviewed","description":"As Liz Truss calls off her sit-down interview, the BBC journalist says in-depth questioning is still needed.","url":"https://www.bbc.co.uk/news/uk-politics-62726849","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/5EB0/production/_126504242_588e29172b43c46264996784bb49f0970d82ff29.jpg","publishedAt":"2022-08-30T15:04:48Z","content":"By Nick RobinsonPresenter, Today programme\r\nAll those people who told me \"She'll never do it. She'd have to be mad. Why would she take the risk?\" turn out to have been right. That's what pretty much … [+6430 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"'If I'd stayed in Afghanistan, I'd be dead'","description":"A year on, thousands of former Afghan government staff who fled the Taliban have nowhere to call home.","url":"https://www.bbc.co.uk/news/world-europe-62679971","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/7FDF/production/_126453723_tophead.jpg","publishedAt":"2022-08-30T15:04:03Z","content":"By Sana Safi &amp; Kawoon KhamooshBBC World Service\r\nA year ago today, Western countries led by the US pulled their troops out of Afghanistan after 20 years, leaving the country in the hands of the T… [+5155 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Sam Rimmer: Fourth suspect bailed after man shot dead in Dingle","description":"A fourth suspect arrested on suspicion of murdering Sam Rimmer is bailed.","url":"https://www.bbc.co.uk/news/uk-england-merseyside-62726920","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/10174/production/_126380956_e62483a5-337e-43f1-a747-ca0530d88f13.jpg","publishedAt":"2022-08-30T15:03:35Z","content":"A fourth suspect arrested on suspicion of murdering a 22-year-old man who was shot in the street has been bailed. \r\nSam Rimmer died in hospital after he was hit in the upper body when shots were fire… [+1108 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Ulster v Exeter pre-season friendly to be shown live on the BBC","description":"Ulster's pre-season friendly against Exeter Chiefs on Friday evening will be shown live on BBC Two Northern Ireland and BBC iPlayer.","url":"https://www.bbc.co.uk/sport/rugby-union/62724311","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/12871/production/_126498857_baloucounewoodburn.png","publishedAt":"2022-08-30T14:59:45Z","content":"Ulster and Exeter last met in the pool stage of the 2016-17 Champions Cup\r\n<table><tr><th>Ulster v Exeter</th></tr>\r\n<tr><td>Venue: Kingspan Stadium, Belfast Date: Friday, 2 September Kick-off: 19:00… [+883 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"The Football News Show: How have Arsenal hit the ground running?","description":"Former Arsenal striker Kevin Campbell tells The Football News Show why boss Mikel Arteta has overseen a fine start to the season.","url":"https://www.bbc.co.uk/sport/av/football/62728926","urlToImage":"https://ichef.bbci.co.uk/news/1024/cpsprodpb/7FE4/production/_126504723_p0cx8d5q.jpg","publishedAt":"2022-08-30T14:56:19Z","content":"The Football News Show looks at how Arsenal have hit the ground running and sit at the top of the Premier League table with four wins from four.\r\nFormer Gunners striker Kevin Campbell joins Ben Crouc… [+156 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"Indonesian soldiers accused of killing Papuans","description":"The victims were allegedly murdered and mutilated after being lured with the promise of weapons.","url":"https://www.bbc.co.uk/news/world-asia-62727954","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/16F58/production/_126504049_hi077070239.jpg","publishedAt":"2022-08-30T14:56:03Z","content":"Six Indonesian soldiers have been arrested on suspicion of killing and mutilating four indigenous Papuans.\r\nThe soldiers are accused of luring their victims to a meeting with the promise of selling t… [+1267 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":null,"title":"Worcester Warriors: Players and staff will not be paid this week","description":"Worcester's players and staff are told they will not be paid on time this week as the future of the Premiership club remains uncertain.","url":"https://www.bbc.co.uk/sport/rugby-union/62728730","urlToImage":"https://ichef.bbci.co.uk/live-experience/cps/624/cpsprodpb/346C/production/_126402431_sixways2.jpg","publishedAt":"2022-08-30T14:54:52Z","content":"Worcester Warriors are scheduled to start the new Premiership season on 10 September\r\nWorcester's players and staff have been told they will not be paid on time this week as the future of the Premier… [+2424 chars]"},{"source":{"id":"bbc-news","name":"BBC News"},"author":"https://www.facebook.com/bbcnews","title":"What's the story? Balamory celebrates 20th anniversary","description":"The hit children's TV show Balamory will return to the BBC iPlayer, as the show marks 20 years since it first aired.","url":"https://www.bbc.co.uk/news/uk-scotland-62722966","urlToImage":"https://ichef.bbci.co.uk/news/1024/branded_news/E56A/production/_126503785_hi026138522.jpg","publishedAt":"2022-08-30T14:24:45Z","content":"By Jack EwingBBC Scotland news\r\nA new generation of children will get to know what the story in Balamory is, as the popular TV show returns to the BBC iPlayer.\r\nBalamory aired between 2002 and 2005 w… [+3239 chars]"}]
-
-class NewsResponse {
-  NewsResponse(
-      {this.status, this.totalResults, this.articles, this.code, this.message});
-
-  NewsResponse.fromJson(dynamic json) {
-    status = json['status'];
-    code = json['code'];
-    message = json['message'];
-    totalResults = json['totalResults'];
-    if (json['articles'] != null) {
-      articles = [];
-      json['articles'].forEach((v) {
-        articles?.add(Articles.fromJson(v));
-      });
-    }
-  }
-
-  String? status;
-  int? totalResults;
-  String? code;
-  String? message;
-  List<Articles>? articles;
-
-  Map<String, dynamic> toJson() {
-    final map = <String, dynamic>{};
-    map['status'] = status;
-    map['totalResults'] = totalResults;
-    if (articles != null) {
-      map['articles'] = articles?.map((v) => v.toJson()).toList();
-    }
-    return map;
-  }
-}
-
-/// source : {"id":"bbc-news","name":"BBC News"}
-/// author : "https://www.facebook.com/bbcnews"
-/// title : "Rent cap: Consultation measure for social housing in England announced"
-/// description : "The government says it wants to avoid rents rising up to 11% for some of England's poorest families."
-/// url : "https://www.bbc.co.uk/news/uk-62732367"
-/// urlToImage : "https://ichef.bbci.co.uk/news/1024/branded_news/DB59/production/_126535165_gettyimages-90678832.jpg"
-/// publishedAt : "2022-08-30T23:02:29Z"
-/// content : "By Andre Rhoden-PaulBBC News\r\nPeople living in social housing in England could see rent increases capped below inflation next year, the government has said. \r\nIt has launched a consultation on plans … [+2610 chars]"
-
-class Articles {
-  Articles({
-    this.source,
-    this.author,
-    this.title,
-    this.description,
-    this.url,
-    this.urlToImage,
-    this.publishedAt,
-    this.content,
-  });
-
-  Articles.fromJson(dynamic json) {
-    source = json['source'] != null ? Source.fromJson(json['source']) : null;
-    author = json['author'];
-    title = json['title'];
-    description = json['description'];
-    url = json['url'];
-    urlToImage = json['urlToImage'];
-    publishedAt = json['publishedAt'];
-    content = json['content'];
-  }
-
-  Source? source;
-  String? author;
-  String? title;
-  String? description;
-  String? url;
-  String? urlToImage;
-  String? publishedAt;
-  String? content;
-
-  Map<String, dynamic> toJson() {
-    final map = <String, dynamic>{};
-    if (source != null) {
-      map['source'] = source?.toJson();
-    }
-    map['author'] = author;
-    map['title'] = title;
-    map['description'] = description;
-    map['url'] = url;
-    map['urlToImage'] = urlToImage;
-    map['publishedAt'] = publishedAt;
-    map['content'] = content;
-    return map;
-  }
-}
-
-/// id : "bbc-news"
-/// name : "BBC News"
