+++ lib/model/NewsResponse.dart
@@ -1,0 +1,82 @@
+import 'package:news_app/model/SoursesRespons.dart';
+
+class NewsResponse {
+  NewsResponse(
+      {this.status, this.totalResults, this.articles, this.code, this.message});
+
+  NewsResponse.fromJson(dynamic json) {
+    status = json['status'];
+    code = json['code'];
+    message = json['message'];
+    totalResults = json['totalResults'];
+    if (json['articles'] != null) {
+      articles = [];
+      json['articles'].forEach((v) {
+        articles?.add(Articles.fromJson(v));
+      });
+    }
+  }
+
+  String? status;
+  int? totalResults;
+  String? code;
+  String? message;
+  List<Articles>? articles;
+
+  Map<String, dynamic> toJson() {
+    final map = <String, dynamic>{};
+    map['status'] = status;
+    map['totalResults'] = totalResults;
+    if (articles != null) {
+      map['articles'] = articles?.map((v) => v.toJson()).toList();
+    }
+    return map;
+  }
+}
+class Articles {
+  Articles({
+    this.source,
+    this.author,
+    this.title,
+    this.description,
+    this.url,
+    this.urlToImage,
+    this.publishedAt,
+    this.content,
+  });
+
+  Articles.fromJson(dynamic json) {
+    source = json['source'] != null ? Source.fromJson(json['source']) : null;
+    author = json['author'];
+    title = json['title'];
+    description = json['description'];
+    url = json['url'];
+    urlToImage = json['urlToImage'];
+    publishedAt = json['publishedAt'];
+    content = json['content'];
+  }
+
+  Source? source;
+  String? author;
+  String? title;
+  String? description;
+  String? url;
+  String? urlToImage;
+  String? publishedAt;
+  String? content;
+
+  Map<String, dynamic> toJson() {
+    final map = <String, dynamic>{};
+    if (source != null) {
+      map['source'] = source?.toJson();
+    }
+    map['author'] = author;
+    map['title'] = title;
+    map['description'] = description;
+    map['url'] = url;
+    map['urlToImage'] = urlToImage;
+    map['publishedAt'] = publishedAt;
+    map['content'] = content;
+    return map;
+  }
+}
