+++ lib/model/SoursesRespons.dart
@@ -1,0 +1,72 @@
+
+class SoursesRespons {
+  String? status;
+  String? code;
+  String? message;
+  List<Source>? sources;
+
+  SoursesRespons({this.status, this.sources, this.code, this.message});
+
+  SoursesRespons.fromJson(dynamic json) {
+    status = json['status'];
+    code = json['code'];
+    message = json['message'];
+    if (json['sources'] != null) {
+      sources = [];
+      json['sources'].forEach((v) {
+        sources?.add(Source.fromJson(v));
+      });
+    }
+  }
+
+  Map<String, dynamic> toJson() {
+    final map = <String, dynamic>{};
+    map['status'] = status;
+    if (sources != null) {
+      map['sources'] = sources?.map((v) => v.toJson()).toList();
+    }
+    return map;
+  }
+}
+
+class Source {
+  Source({
+    this.id,
+    this.name,
+    this.description,
+    this.url,
+    this.category,
+    this.language,
+    this.country,
+  });
+
+  Source.fromJson(dynamic json) {
+    id = json['id'];
+    name = json['name'];
+    description = json['description'];
+    url = json['url'];
+    category = json['category'];
+    language = json['language'];
+    country = json['country'];
+  }
+
+  String? id;
+  String? name;
+  String? description;
+  String? url;
+  String? category;
+  String? language;
+  String? country;
+
+  Map<String, dynamic> toJson() {
+    final map = <String, dynamic>{};
+    map['id'] = id;
+    map['name'] = name;
+    map['description'] = description;
+    map['url'] = url;
+    map['category'] = category;
+    map['language'] = language;
+    map['country'] = country;
+    return map;
+  }
+}
