+++ lib/views/constant/constant.dart
@@ -1,0 +1,2 @@
+const API_KEY = '1a857ebff2674b71929900fabb8e3f56';
+const BASE_URL = 'newsapi.org';
