+++ lib/views/news/news_details/homeScreen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:news_app/presentation_layer/categories/categories.dart';
-import 'package:news_app/presentation_layer/categories/categoryModel.dart';
-import 'package:news_app/presentation_layer/homeSideMinue.dart';
-import 'package:news_app/presentation_layer/news/newsFragment.dart';
-import 'package:news_app/presentation_layer/news/newsSearch.dart';
-import 'package:news_app/presentation_layer/settings/settings.dart';
+import 'package:news_app/views/categories/categories.dart';
+import 'package:news_app/views/categories/categoryModel.dart';
+import 'package:news_app/views/drawer/homeSideMinue.dart';
+import 'package:news_app/views/news/news_details/newsFragment.dart';
+import 'package:news_app/views/news/search/newsSearch.dart';
+import 'package:news_app/views/settings/settings.dart';
 
 class HomeScreen extends StatefulWidget {
   static const String routeName = 'home screen';
