+++ lib/views/news/news_details/newsFragment.dart
@@ -1,19 +1,25 @@
 import 'package:flutter/material.dart';
-import 'package:news_app/api/api_manager.dart';
-import 'package:news_app/presentation_layer/categories/categoryModel.dart';
-import 'package:news_app/presentation_layer/news/Sources_tabs.dart';
+import 'package:news_app/model/SoursesRespons.dart';
+import 'package:news_app/repository/NewsRepositpry.dart';
+import 'package:news_app/view_models/news_viewModel.dart';
+import 'package:news_app/views/categories/categoryModel.dart';
+import 'package:news_app/views/news/source/Sources_tabs.dart';
 
 class NewsFragment extends StatelessWidget {
   BuildCategory category;
 
   NewsFragment(this.category);
 
+  NewsDetailsViewModel newsDetailsViewModel = NewsDetailsViewModel(
+    repository: NewsRepository(),
+  );
+
   @override
   Widget build(BuildContext context) {
     return Column(
       children: [
-        FutureBuilder(
-            future: ApiManager.getNewsSource(category.id!),
+        FutureBuilder<SoursesRespons>(
+            future: newsDetailsViewModel.fetchNewsSource(category.id!),
             builder: (buildContext, snapShot) {
               if (snapShot.hasError) {
                 return Center(child: Text('${snapShot.error.toString()}'));
@@ -28,14 +34,6 @@ class NewsFragment extends StatelessWidget {
               }
               var sources = data?.sources;
               return SourcesTabs(sources);
-              /*
-               ListView.builder(
-                     itemBuilder: (_,index){
-                       return Text((sources![index].name) ?? '') ;
-                     },
-                 itemCount: sources?.length,
-                 ),
-                */
             }),
       ],
     );
