+++ lib/views/news/search/newsList.dart
@@ -1,19 +1,24 @@
 import 'package:flutter/material.dart';
-import 'package:news_app/api/api_manager.dart';
-import 'package:news_app/api/model/NewsResponse.dart';
-import 'package:news_app/api/model/SoursesRespons.dart';
-import 'package:news_app/presentation_layer/news/newsWidget.dart';
+import 'package:news_app/model/NewsResponse.dart';
+import 'package:news_app/model/SoursesRespons.dart';
+import 'package:news_app/repository/NewsRepositpry.dart';
+import 'package:news_app/view_models/news_viewModel.dart';
+import 'package:news_app/views/news/search/newsWidget.dart';
 
 class NewsList extends StatelessWidget {
   Source source;
 
   NewsList(this.source);
 
+  NewsDetailsViewModel newsDetailsViewModel = NewsDetailsViewModel(
+    repository: NewsRepository(),
+  );
+
   @override
   Widget build(BuildContext context) {
     return Container(
       child: FutureBuilder<NewsResponse>(
-        future: ApiManager.getNewsBySourceId(sourceId: source.id!),
+        future: newsDetailsViewModel.getNewsBySourceId(sourceId: source.id!),
         builder: (buildContext, snapShot) {
           if (snapShot.hasError) {
             return Center(child: Text('${snapShot.error.toString()}'));
