+++ lib/views/news/search/newsWidget.dart
@@ -1,8 +1,8 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:news_app/api/model/NewsResponse.dart';
-import 'package:news_app/presentation_layer/news/my_date_utiles.dart';
-import 'package:news_app/presentation_layer/news/newsDetails.dart';
+import 'package:news_app/model/NewsResponse.dart';
+import 'package:news_app/views/constant/my_date_utiles.dart';
+import 'package:news_app/views/news/news_details/newsDetails.dart';
 
 class NewsWidget extends StatelessWidget {
   Articles articles;
